+++ Framework/System/Time.cpp
@@ -20,7 +20,7 @@ void Time::Update()
 	//Limit framerate
 	float worktime = (current - m_previous) / (float)m_frequency;
 	float sleeptime = std::max(m_frametime - worktime, 0.0f);
-	Sleep(sleeptime * 1000.0f);
+	Sleep(static_cast<DWORD>(sleeptime * 1000.0f));
 
 	m_elapsed = worktime + sleeptime;
